tests/wpt/tests/css/CSS2/floats/zero-space-between-floats-003.html

<!DOCTYPE html>
<title>Zero-width new formatting context with clearance</title>
<link rel="author" title="Morten Stenshorne" href="mstensho@chromium.org">
<link rel="help" href="https://www.w3.org/TR/CSS22/visuren.html#float-position" title="9.5.1 Positioning the float: the 'float' property">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<div id="container" style="position:relative; width:200px;">
  <div style="float:left; width:100px; height:100px;"></div>
  <div style="float:right; width:100px; height:100px;"></div>
  <div style="float:right; width:100%; height:100px;"></div>
  <div data-offset-x="0" data-offset-y="100" style="overflow:hidden; clear:left; width:0;"></div>
</div>
<script>
  checkLayout("#container");
</script>
